[python] Add CLI command to run tests
